compute actual sample rate for profile export
In cases where due to either an excessive number of goroutines, or excessive
load on the system, we are not able to sample at the given rate, we can compute
the actual rate we were able to sample and use that when exporting the profile
so that timing information in the profile is more accurate and sensible.

With this code in place, a function which is on the stack for the entirety of a
27 second profile, will always display 27s when viewed, whereas previously it
could display less time.